archive/old_files/requirements.txt
# Core dependencies with specific versions
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==20.1.0

# Supabase and related dependencies
supabase==1.0.3
httpx==0.24.0
httpcore==0.15.0
pydantic==1.10.7
postgrest-py==0.10.6
python-dotenv==1.0.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# Google Cloud dependencies
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# File processing dependencies
pypdf2==3.0.1
pillow==10.0.0
xlrd==2.0.1
openpyxl==3.1.2

# Database dependencies
psycopg2-binary==2.9.7

# Other dependencies
requests==2.31.0
requests-oauthlib==1.3.1
email-validator==2.0.0
pytest==7.3.1
pytest-flask==1.2.0
pytest-cov==4.1.0
aiohttp==3.9.3
python-multipart==0.0.6
python-magic==0.4.27
boto3==1.28.36
botocore==1.31.36
typing_extensions==4.7.1
urllib3==1.26.18
websockets==11.0.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
setuptools==68.2.2
wheel==0.41.2
pip==23.2.1

# Added from the code block
fastapi==0.95.1
uvicorn==0.22.0
uuid==1.30
pytest-asyncio==0.21.0 